Continue main map page to vue component
