+++ package.json
@@ -9,6 +9,7 @@
   },
   "dependencies": {
     "@fortawesome/fontawesome-free": "^5.15.3",
+    "@turf/distance": "^6.5.0",
     "c3": "^0.7.20",
     "core-js": "^3.6.5",
     "d3": "^5.16.0",
@@ -16,6 +17,8 @@
     "font-awesome": "^4.7.0",
     "jquery": "^2.2.4",
     "jquery-ui": "^1.10.5",
+    "leaflet": "^1.9.3",
+    "polyline-encoded": "0.0.9",
     "sass-loader": "^10.1.1",
     "vue": "^2.6.11",
     "vue-analytics": "^5.22.1",
