+++ public/index.html
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<html lang="">
+<html lang="en-US">
 
 <head>
   <meta charset="utf-8" />
@@ -24,6 +24,9 @@
   <meta property="og:title" content="Jeffrey Piccirillo" />
   <meta property="og:site_name" content="Personal Projects and Portfolio | Jeffrey Piccirillo" />
   <title>Jeffrey Piccirillo</title>
+  <link rel="stylesheet" href="https://unpkg.com/leaflet@1.7.1/dist/leaflet.css"
+    integrity="sha512-xodZBNTC5n17Xt2atTPuE1HxjVMSvLVW9ocqUKLsCC5CXdbqCmblAshOMAS6/keqq/sMZMZ19scR4PsZChSR7A=="
+    crossorigin="" />
 </head>
 
 <body>
